refactored rms_norm and test_emit_verilog_norm test
