+++ machop/mase_components/norm/rtl/norm.sv
@@ -105,8 +105,8 @@ if (LAYER_NORM || INSTANCE_NORM || GROUP_NORM) begin : group_norm
         .COMPUTE_DIM0(COMPUTE_DIM0),
         .COMPUTE_DIM1(COMPUTE_DIM1),
         .GROUP_CHANNELS(NORM_CHANNELS),
-        .IN_FRAC_WIDTH(IN_FRAC_WIDTH),
         .IN_WIDTH(IN_WIDTH),
+        .IN_FRAC_WIDTH(IN_FRAC_WIDTH),
         .OUT_WIDTH(OUT_WIDTH),
         .OUT_FRAC_WIDTH(OUT_FRAC_WIDTH),
         .ISQRT_LUT_MEMFILE(ISQRT_LUT_MEMFILE)
@@ -123,26 +123,27 @@ if (LAYER_NORM || INSTANCE_NORM || GROUP_NORM) begin : group_norm
 
 end else if (RMS_NORM) begin : rms_norm
 
-    // rms_norm_2d #(
-    //     .TOTAL_DIM0(TOTAL_DIM0),
-    //     .TOTAL_DIM1(TOTAL_DIM1),
-    //     .COMPUTE_DIM0(COMPUTE_DIM0),
-    //     .COMPUTE_DIM1(COMPUTE_DIM1),
-    //     .CHANNELS(CHANNELS),
-    //     .IN_FRAC_WIDTH(IN_FRAC_WIDTH),
-    //     .IN_WIDTH(IN_WIDTH),
-    //     .OUT_WIDTH(OUT_WIDTH),
-    //     .OUT_FRAC_WIDTH(OUT_FRAC_WIDTH)
-    // ) rms_norm_inst (
-    //     .clk(clk),
-    //     .rst(rst),
-    //     .in_data(data_in_0),
-    //     .in_valid(data_in_0_valid),
-    //     .in_ready(data_in_0_ready),
-    //     .out_data(data_out_0),
-    //     .out_valid(data_out_0_valid),
-    //     .out_ready(data_out_0_ready)
-    // );
+    rms_norm_2d #(
+        .TOTAL_DIM0(TOTAL_DIM0),
+        .TOTAL_DIM1(TOTAL_DIM1),
+        .COMPUTE_DIM0(COMPUTE_DIM0),
+        .COMPUTE_DIM1(COMPUTE_DIM1),
+        .CHANNELS(CHANNELS),
+        .IN_WIDTH(IN_WIDTH),
+        .IN_FRAC_WIDTH(IN_FRAC_WIDTH),
+        .OUT_WIDTH(OUT_WIDTH),
+        .OUT_FRAC_WIDTH(OUT_FRAC_WIDTH),
+        .ISQRT_LUT_MEMFILE(ISQRT_LUT_MEMFILE)
+    ) rms_norm_inst (
+        .clk(clk),
+        .rst(rst),
+        .in_data(data_in_0),
+        .in_valid(data_in_0_valid),
+        .in_ready(data_in_0_ready),
+        .out_data(data_out_0),
+        .out_valid(data_out_0_valid),
+        .out_ready(data_out_0_ready)
+    );
 
 end
